Write showSuperAdmins -Write API call -Write action -Write store -Set dispatcher correctly
KNOW BUG: the first time you log in with a super admin account and you go in
the show super admin list you won't see any entries, and I don't know why.
This will be fixed in the next commits.
IMPORTANT: I'm using a API that doesn't exist for the moment. This task is WIP.
For more news see issue #130.